+++ Prime/Prime.h
@@ -7,5 +7,5 @@ public:
 	bool isPrime(int num);
 	int HighestPrime(int num);
 	int SumOfPrimes(int num);
-	int HighestPrimeFactor(int num);
+	int HighestPrimeFactor(long long num);
 };
